+++ R/options.R
@@ -188,6 +188,8 @@
 #' R_FUTURE_OUTPUT_WINDOWS_REENCODE
 #' future.journal
 #' R_FUTURE_JOURNAL
+#' R_FUTURE_GLOBALS_OBJECTSIZE_METHOD
+#' future.globals.objectSize.method
 #'
 #' @name future.options
 NULL
@@ -358,5 +360,12 @@ update_package_options <- function(debug = FALSE) {
   update_package_option("future.globals.keepWhere", mode = "logical", debug = debug)
 
   ## future 1.34.0:
+  update_package_option("future.globals.objectSize.method", mode = "character", debug = debug)
+  if (is.null(getOption("future.globals.objectSize.method"))) {
+   if (exists("serializedSize", envir = getNamespace("parallelly"), inherits = FALSE)) {
+      options(future.globals.objectSize.method = "serializedSize")
+    }
+  }
+
   update_package_option("future.plan.cleanup.legacy", mode = "logical", debug = debug)
 }