+++ R/utils-objectSize.R
@@ -5,6 +5,19 @@ objectSize <- function(x, depth = 3L, enclosure = getOption("future.globals.obje
   # Nothing to do?
   if (isNamespace(x)) return(0)
   if (depth <= 0) return(0)
+
+  method <- getOption("future.globals.objectSize.method", "objectSize")
+  if (method == "serializedSize") {
+    ns <- getNamespace("parallelly")
+    if (!exists("serializedSize", envir = ns, inherits = FALSE)) {
+      stop("Option 'future.globals.objectSize.method' supports \"serializedSize\" only for parallelly (>= 1.38.0)")
+    }
+    serializedSize <- get("serializedSize", envir = ns, inherits = FALSE)
+    size <- serializedSize(x)
+    return(size)
+  } else if (method != "objectSize") {
+    stop(paste("Unknown value on option 'future.globals.objectSize.method':", sQuote(method)))
+  }
   
   if (!is.list(x) && !is.environment(x)) {
     size <- unclass(object.size(x))