+++ tests/multisession.R
@@ -108,44 +108,6 @@ for (cores in 1:availCores) {
 } ## for (cores ...)
 
 
-message("*** multisession() - too large globals ...")
-ooptsT <- options(future.globals.maxSize = object.size(1:1014))
-
-limit <- getOption("future.globals.maxSize")
-cat(sprintf("Max total size of globals: %g bytes\n", limit))
-
-for (workers in unique(c(1L, availableCores()))) {
-  ## Speed up CRAN checks: Skip on CRAN Windows 32-bit
-  if (!fullTest && isWin32) next
-  
-  message("Max number of sessions: ", workers)
-
-  ## A large object
-  a <- 1:1014
-  yTruth <- sum(a)
-  size <- object.size(a)
-  cat(sprintf("a: %g bytes\n", size))
-  f <- multisession({ sum(a) }, globals = TRUE, workers = workers)
-  print(f)
-  rm(list = "a")
-  v <- value(f)
-  print(v)
-  stopifnot(v == yTruth)
-
-
-  ## A too large object
-  a <- 1:1015
-  yTruth <- sum(a)
-  size <- object.size(a)
-  cat(sprintf("a: %g bytes\n", size))
-  res <- try(f <- multisession({ sum(a) }, globals = TRUE, workers = workers), silent = TRUE)
-  rm(list = "a")
-  stopifnot(inherits(res, "try-error"))
-} ## for (workers in ...)
-
-## Undo options changed in this test
-options(ooptsT)
-
 message("*** multisession() - too large globals ... DONE")
 
 message("*** multisession(..., workers = 1L) ...")